% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distance function.R
\name{get_distance}
\alias{get_distance}
\title{Return the Basic Facts About a US College}
\usage{
get_distance(hometown, college, unit = "miles")
}
\arguments{
\item{hometown}{Location to get distance from}

\item{college}{The name of the the school of interest.}

\item{units}{The preferred units for reporting the distance.}
}
\description{
Return the Basic Facts About a US College
}
\examples{
get_info("University of Virginia")
}

