% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info function.R
\name{get_info}
\alias{get_info}
\title{Return the Basic Facts About a US College}
\usage{
get_info(college)
}
\arguments{
\item{college}{The name of the the school of interest.}
}
\description{
Return the Basic Facts About a US College
}
\examples{
get_info("University of Virginia")
}

